sweet alerts añadidos, sección atracciones completa

+++ src/app/pages/attractions-pages/attraction-details/attraction-details.component.ts
@@ -2,6 +2,7 @@ import { Component, inject, Input } from '@angular/core';
 import { IAttraction } from '../../../interfaces/iattraction.interface';
 import { AttractionsService } from '../../../services/attractions.service';
 import { Router, RouterModule } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-attraction-details',
@@ -28,12 +29,15 @@ export class AttractionDetailsComponent {
   }
 
   async onClick() {
-
-    try {
-      await this.attractionsService.deleteById(this.attraction!.id);
-      this.router.navigateByUrl('attractions');
-    } catch (error) {
-      console.log(error);
+    const confirm = await Swal.fire({ title: 'Confirmar', text: 'Estás a punto de borrar tanto los datos como la atracción en sí misma', icon: 'warning', confirmButtonText: 'Confirmar', showCancelButton: true });
+    if (confirm.isConfirmed) {
+      try {
+        await this.attractionsService.deleteById(this.attraction!.id);
+        Swal.fire('Borrado', 'La atracción ha sido eliminada por completo y para siempre', 'success');
+        this.router.navigateByUrl('attractions');
+      } catch (error) {
+        console.log(error);
+      }
     }
   }
